split up giant client spec using shared contexts and examples
